n_layer

+++ ODIN/densenet.py
@@ -91,9 +91,9 @@ class DenseNet3(nn.Module):
                  reduction=0.5, bottleneck=True, dropRate=0.0):
         super(DenseNet3, self).__init__()
         in_planes = 2 * growth_rate
-        n = (depth - 4) / 3
+        n = int((depth - 4) / 3)
         if bottleneck == True:
-            n = n/2
+            n = int(n/2)
             block = BottleneckBlock
         else:
             block = BasicBlock
